Separate Fee and Fee Storage class from ICX Storage
- Separates unit test regarding Fee and Deposit from ICX Storage
