+++ iconservice/fee/deposit.py
@@ -37,7 +37,8 @@ class Deposit(object):
 
     def __init__(self, deposit_id: bytes = None, score_address: 'Address' = None, sender: 'Address' = None,
                  deposit_amount: int = 0, deposit_used: int = 0, created: int = 0, expires: int = 0,
-                 virtual_step_issued: int = 0, virtual_step_used: int = 0, prev_id: bytes = None, next_id: bytes = None):
+                 virtual_step_issued: int = 0, virtual_step_used: int = 0, prev_id: bytes = None,
+                 next_id: bytes = None):
         # deposit id, should be tx hash of deposit transaction
         self.id = deposit_id
         # target SCORE address
@@ -96,7 +97,8 @@ class Deposit(object):
                                  self.created.to_bytes(DEFAULT_BYTE_SIZE, DATA_BYTE_ORDER),
                                  self.expires.to_bytes(DEFAULT_BYTE_SIZE, DATA_BYTE_ORDER),
                                  self.virtual_step_issued.to_bytes(DEFAULT_BYTE_SIZE, DATA_BYTE_ORDER),
-                                 self.virtual_step_used.to_bytes(DEFAULT_BYTE_SIZE, DATA_BYTE_ORDER), self.prev_id, self.next_id)
+                                 self.virtual_step_used.to_bytes(DEFAULT_BYTE_SIZE, DATA_BYTE_ORDER), self.prev_id,
+                                 self.next_id)
 
     def to_dict(self, casing: Optional = None) -> dict:
         """Returns properties as dict
@@ -121,16 +123,16 @@ class Deposit(object):
         :param other: (Deposit)
         """
         return isinstance(other, Deposit) \
-            and self.score_address == other.score_address \
-            and self.sender == other.sender \
-            and self.deposit_amount == other.deposit_amount \
-            and self.deposit_used == other.deposit_used \
-            and self.created == other.created \
-            and self.expires == other.expires \
-            and self.virtual_step_issued == other.virtual_step_issued \
-            and self.virtual_step_used == other.virtual_step_used \
-            and self.prev_id == other.prev_id \
-            and self.next_id == other.next_id
+               and self.score_address == other.score_address \
+               and self.sender == other.sender \
+               and self.deposit_amount == other.deposit_amount \
+               and self.deposit_used == other.deposit_used \
+               and self.created == other.created \
+               and self.expires == other.expires \
+               and self.virtual_step_issued == other.virtual_step_issued \
+               and self.virtual_step_used == other.virtual_step_used \
+               and self.prev_id == other.prev_id \
+               and self.next_id == other.next_id
 
     def __ne__(self, other) -> bool:
         """operator != overriding
@@ -138,6 +140,3 @@ class Deposit(object):
         :param other: (Deposit)
         """
         return not self.__eq__(other)
-
-
-
